reorganización de las rutas de archivo
